requirements.txt
wheel
torch
torchvision
torchaudio
lightning
hydra-core
einops
jaxtyping
beartype
Pillow
wandb
dacite
black
ruff
timm
matplotlib
scipy
plyfile
click
git+https://github.com/ChristophReich1996/Optical-Flow-Visualization-PyTorch
